Title: 특정 구간의 원소 평균값, Time: 139ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240525/특정 구간의 원소 평균값/elemental-mean-value-for-a-particular-interval.java
@@ -21,7 +21,7 @@ public class Main {
                 for(int k=i; k<j; k++){
                     sum += arr[k];
                 }
-                int avg = sum/(j-i);
+                double avg = sum/(j-i);
 
                 boolean check = false;
                 for(int k=i; k<j; k++){
